examples/chatgpt/pyproject.toml
[project]
authors = [
    { name = "Laurent Fainsin", email = "laurent@lagon.tech" },
]
dependencies = [
    "environs>=14.1.0",
    "finegrain",
    "pillow>=11.1.0",
    "pydantic>=2.10.5",
    "quart>=0.20.0",
]
description = "Finegrain API ChatGPT bridge"
name = "chatgpt-bridge"
readme = "README.md"
requires-python = ">= 3.12"
version = "1.0.0"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.396",
    "ruff>=0.11.0",
    "typos>=1.30.0",
]
package = true

[tool.uv.sources]
finegrain = { git = "https://github.com/finegrain-ai/finegrain-python.git", subdirectory = "finegrain" }

[build-system]
build-backend = "hatchling.build"
requires = ["hatch-vcs", "hatchling"]

[tool.hatch.version]
fallback-version = '0.0.0'
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/chatgpt_bridge"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "C901",  # is too complex
    "E731",  # do-not-assign-lambda
    "EM101", # exception must not use a string literal
    "EM102", # f-string literal in exception message
    "G004",  # f-string literal in logging message
    "N812",  # imported as non-lowercase
    "S101",  # use of assert
    "S311",  # non secure cryptographic random
]
select = [
    "A",   # flake8-builtins
    "B",   # flake8-bugbear
    "E",   # pycodestyle errors
    "EM",  # flake8-errmsg
    "F",   # Pyflakes
    "G",   # flake8-logging-format
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PTH", # flake8-use-pathlib
    "RUF", # ruff
    "S",   # flake8-bandit
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.pyright]
pythonVersion = "3.12"
reportMissingImports = true
reportMissingTypeStubs = false
reportPrivateUsage = false
reportUntypedFunctionDecorator = false
